+++ Addon/ElvUI_mMediaTag/core/changelog/3.52.lua
@@ -1,0 +1,8 @@
+mMT.Changelog[352] = {
+	DATE = "15.08.2024",
+	FIX = {
+		"[Portrait]: fixed a Bug with Class icons and the recent Version of JiberishUI Icons",
+        "[Portrait]: prevent nil error on Profile switch",
+        "[System]: custom/ example Datatext menu entry is back again",
+	},
+}